# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * CogniLearn Adaptive Training Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateSessionBodyGameType = typeof CreateSessionBodyGameType[keyof typeof CreateSessionBodyGameType];


export const CreateSessionBodyGameType = {
  attention_response: 'attention_response',
  continuous_performance: 'continuous_performance',
  task_switching: 'task_switching',
  letter_sound: 'letter_sound',
  word_formation: 'word_formation',
  visual_tracking: 'visual_tracking',
  highlighted_reading: 'highlighted_reading',
  social_scenario: 'social_scenario',
  emotion_recognition: 'emotion_recognition',
  sensory_filtering: 'sensory_filtering',
  letter_recognition: 'letter_recognition',
  pattern_recognition: 'pattern_recognition',
  sequencing_task: 'sequencing_task',
  audio_word_match: 'audio_word_match',
  speech_echo: 'speech_echo',
  mirror_match: 'mirror_match',
} as const;

export interface CreateSessionBody {
  gameType: CreateSessionBodyGameType;
  userId?: number | null;
  /**
     * @minimum 1
     * @maximum 5
     */
  difficulty?: number;
}

export type UpdateSessionBodyAdaptiveFlags = {
  lowAttention?: boolean;
  readingDifficulty?: boolean;
  sensoryOverload?: boolean;
  socialDifficulty?: boolean;
};

export interface UpdateSessionBody {
  score?: number;
  accuracy?: number;
  reactionTimeMs?: number;
  wrongClicks?: number;
  missedSignals?: number;
  attentionDuration?: number;
  hesitationCount?: number;
  completedAt?: string;
  adaptiveFlags?: UpdateSessionBodyAdaptiveFlags;
}

export type SessionAdaptiveFlags = { [key: string]: unknown } | null;

export interface Session {
  id: number;
  gameType: string;
  difficulty: number;
  score?: number;
  accuracy?: number;
  reactionTimeMs?: number;
  wrongClicks?: number;
  missedSignals?: number;
  attentionDuration?: number;
  hesitationCount?: number;
  startedAt: string;
  completedAt?: string | null;
  adaptiveFlags?: SessionAdaptiveFlags;
}

export type ProgressSummarySkillLevels = {
  attention?: number;
  memory?: number;
  socialCognition?: number;
  readingPhonics?: number;
  sensoryProcessing?: number;
};

export interface ProgressSummary {
  totalSessions: number;
  totalMinutes: number;
  averageAccuracy: number;
  averageReactionTimeMs?: number;
  currentStreak: number;
  skillLevels: ProgressSummarySkillLevels;
}

export type GameProgressTrendDirection = typeof GameProgressTrendDirection[keyof typeof GameProgressTrendDirection];


export const GameProgressTrendDirection = {
  improving: 'improving',
  stable: 'stable',
  declining: 'declining',
} as const;

export interface GameProgress {
  gameType: string;
  sessionsPlayed: number;
  averageScore: number;
  averageAccuracy: number;
  bestScore: number;
  currentDifficulty: number;
  trendDirection: GameProgressTrendDirection;
}

export interface ActivityEntry {
  id: number;
  gameType: string;
  score?: number;
  accuracy?: number;
  difficulty?: number;
  completedAt: string;
}

export type AdaptiveSettingsFontSize = typeof AdaptiveSettingsFontSize[keyof typeof AdaptiveSettingsFontSize];


export const AdaptiveSettingsFontSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
  xlarge: 'xlarge',
} as const;

export type AdaptiveSettingsColorTheme = typeof AdaptiveSettingsColorTheme[keyof typeof AdaptiveSettingsColorTheme];


export const AdaptiveSettingsColorTheme = {
  default: 'default',
  high_contrast: 'high_contrast',
  warm: 'warm',
  cool: 'cool',
} as const;

export interface AdaptiveSettings {
  id: number;
  taskDurationSeconds: number;
  fontSize: AdaptiveSettingsFontSize;
  audioEnabled: boolean;
  visualHighlights: boolean;
  simplifiedUi: boolean;
  breakFrequencyMinutes: number;
  colorTheme: AdaptiveSettingsColorTheme;
  /**
     * @minimum 1
     * @maximum 5
     */
  currentDifficulty: number;
}

export type UpdateAdaptiveSettingsBodyFontSize = typeof UpdateAdaptiveSettingsBodyFontSize[keyof typeof UpdateAdaptiveSettingsBodyFontSize];


export const UpdateAdaptiveSettingsBodyFontSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
  xlarge: 'xlarge',
} as const;

export type UpdateAdaptiveSettingsBodyColorTheme = typeof UpdateAdaptiveSettingsBodyColorTheme[keyof typeof UpdateAdaptiveSettingsBodyColorTheme];


export const UpdateAdaptiveSettingsBodyColorTheme = {
  default: 'default',
  high_contrast: 'high_contrast',
  warm: 'warm',
  cool: 'cool',
} as const;

export interface UpdateAdaptiveSettingsBody {
  taskDurationSeconds?: number;
  fontSize?: UpdateAdaptiveSettingsBodyFontSize;
  audioEnabled?: boolean;
  visualHighlights?: boolean;
  simplifiedUi?: boolean;
  breakFrequencyMinutes?: number;
  colorTheme?: UpdateAdaptiveSettingsBodyColorTheme;
  /**
     * @minimum 1
     * @maximum 5
     */
  currentDifficulty?: number;
}

export type GameRecommendationPriority = typeof GameRecommendationPriority[keyof typeof GameRecommendationPriority];


export const GameRecommendationPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface GameRecommendation {
  gameType: string;
  reason: string;
  priority: GameRecommendationPriority;
  suggestedDifficulty: number;
}

export type ListSessionsParams = {
gameType?: string;
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};