# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CogniLearn Adaptive Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adaptive/settings:
    get:
      operationId: getAdaptiveSettings
      tags: [adaptive]
      summary: Get current adaptive settings for a learner
      responses:
        "200":
          description: Adaptive settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveSettings"
    patch:
      operationId: updateAdaptiveSettings
      tags: [adaptive]
      summary: Update adaptive settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdaptiveSettingsBody"
      responses:
        "200":
          description: Updated adaptive settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveSettings"
...
components:
  schemas:
...
    AdaptiveSettings:
      type: object
      properties:
        id:
          type: integer
        taskDurationSeconds:
          type: integer
        fontSize:
          type: string
          enum: [small, medium, large, xlarge]
        audioEnabled:
          type: boolean
        visualHighlights:
          type: boolean
        simplifiedUi:
          type: boolean
        breakFrequencyMinutes:
          type: integer
        colorTheme:
          type: string
          enum: [default, high_contrast, warm, cool]
        currentDifficulty:
          type: integer
          minimum: 1
          maximum: 5
      required:
        - id
        - taskDurationSeconds
        - fontSize
        - audioEnabled
        - visualHighlights
        - simplifiedUi
        - breakFrequencyMinutes
        - colorTheme
        - currentDifficulty
    UpdateAdaptiveSettingsBody:
      type: object
      properties:
        taskDurationSeconds:
          type: integer
        fontSize:
          type: string
          enum: [small, medium, large, xlarge]
        audioEnabled:
          type: boolean
        visualHighlights:
          type: boolean
        simplifiedUi:
          type: boolean
        breakFrequencyMinutes:
          type: integer
        colorTheme:
          type: string
          enum: [default, high_contrast, warm, cool]
        currentDifficulty:
          type: integer
          minimum: 1
          maximum: 5
...