# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * CogniLearn Adaptive Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Create a new game session
 */
export const createSessionBodyDifficultyMax = 5;



export const CreateSessionBody = zod.object({
  "gameType": zod.enum(['attention_response', 'continuous_performance', 'task_switching', 'letter_sound', 'word_formation', 'visual_tracking', 'highlighted_reading', 'social_scenario', 'emotion_recognition', 'sensory_filtering', 'letter_recognition', 'pattern_recognition', 'sequencing_task', 'audio_word_match', 'speech_echo', 'mirror_match']),
  "userId": zod.number().nullish(),
  "difficulty": zod.number().min(1).max(createSessionBodyDifficultyMax).optional()
})


/**
 * @summary List all sessions
 */
export const listSessionsQueryLimitDefault = 20;

export const ListSessionsQueryParams = zod.object({
  "gameType": zod.coerce.string().optional(),
  "limit": zod.coerce.number().default(listSessionsQueryLimitDefault)
})

export const ListSessionsResponseItem = zod.object({
  "id": zod.number(),
  "gameType": zod.string(),
  "difficulty": zod.number(),
  "score": zod.number().optional(),
  "accuracy": zod.number().optional(),
  "reactionTimeMs": zod.number().optional(),
  "wrongClicks": zod.number().optional(),
  "missedSignals": zod.number().optional(),
  "attentionDuration": zod.number().optional(),
  "hesitationCount": zod.number().optional(),
  "startedAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish(),
  "adaptiveFlags": zod.object({

}).passthrough().nullish()
})
export const ListSessionsResponse = zod.array(ListSessionsResponseItem)


/**
 * @summary Get a specific session
 */
export const GetSessionParams = zod.object({
  "id": zod.coerce.number()
})

export const GetSessionResponse = zod.object({
  "id": zod.number(),
  "gameType": zod.string(),
  "difficulty": zod.number(),
  "score": zod.number().optional(),
  "accuracy": zod.number().optional(),
  "reactionTimeMs": zod.number().optional(),
  "wrongClicks": zod.number().optional(),
  "missedSignals": zod.number().optional(),
  "attentionDuration": zod.number().optional(),
  "hesitationCount": zod.number().optional(),
  "startedAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish(),
  "adaptiveFlags": zod.object({

}).passthrough().nullish()
})


/**
 * @summary Update/complete a session
 */
export const UpdateSessionParams = zod.object({
  "id": zod.coerce.number()
})

export const UpdateSessionBody = zod.object({
  "score": zod.number().optional(),
  "accuracy": zod.number().optional(),
  "reactionTimeMs": zod.number().optional(),
  "wrongClicks": zod.number().optional(),
  "missedSignals": zod.number().optional(),
  "attentionDuration": zod.number().optional(),
  "hesitationCount": zod.number().optional(),
  "completedAt": zod.coerce.date().optional(),
  "adaptiveFlags": zod.object({
  "lowAttention": zod.boolean().optional(),
  "readingDifficulty": zod.boolean().optional(),
  "sensoryOverload": zod.boolean().optional(),
  "socialDifficulty": zod.boolean().optional()
}).optional()
})

export const UpdateSessionResponse = zod.object({
  "id": zod.number(),
  "gameType": zod.string(),
  "difficulty": zod.number(),
  "score": zod.number().optional(),
  "accuracy": zod.number().optional(),
  "reactionTimeMs": zod.number().optional(),
  "wrongClicks": zod.number().optional(),
  "missedSignals": zod.number().optional(),
  "attentionDuration": zod.number().optional(),
  "hesitationCount": zod.number().optional(),
  "startedAt": zod.coerce.date(),
  "completedAt": zod.coerce.date().nullish(),
  "adaptiveFlags": zod.object({

}).passthrough().nullish()
})


/**
 * @summary Get overall progress summary
 */
export const GetProgressResponse = zod.object({
  "totalSessions": zod.number(),
  "totalMinutes": zod.number(),
  "averageAccuracy": zod.number(),
  "averageReactionTimeMs": zod.number().optional(),
  "currentStreak": zod.number(),
  "skillLevels": zod.object({
  "attention": zod.number().optional(),
  "memory": zod.number().optional(),
  "socialCognition": zod.number().optional(),
  "readingPhonics": zod.number().optional(),
  "sensoryProcessing": zod.number().optional()
})
})


/**
 * @summary Get progress breakdown by game type
 */
export const GetProgressByGameResponseItem = zod.object({
  "gameType": zod.string(),
  "sessionsPlayed": zod.number(),
  "averageScore": zod.number(),
  "averageAccuracy": zod.number(),
  "bestScore": zod.number(),
  "currentDifficulty": zod.number(),
  "trendDirection": zod.enum(['improving', 'stable', 'declining'])
})
export const GetProgressByGameResponse = zod.array(GetProgressByGameResponseItem)


/**
 * @summary Get recent activity timeline
 */
export const getRecentActivityQueryLimitDefault = 10;

export const GetRecentActivityQueryParams = zod.object({
  "limit": zod.coerce.number().default(getRecentActivityQueryLimitDefault)
})

export const GetRecentActivityResponseItem = zod.object({
  "id": zod.number(),
  "gameType": zod.string(),
  "score": zod.number().optional(),
  "accuracy": zod.number().optional(),
  "difficulty": zod.number().optional(),
  "completedAt": zod.coerce.date()
})
export const GetRecentActivityResponse = zod.array(GetRecentActivityResponseItem)


/**
 * @summary Get current adaptive settings for a learner
 */
export const getAdaptiveSettingsResponseCurrentDifficultyMax = 5;



export const GetAdaptiveSettingsResponse = zod.object({
  "id": zod.number(),
  "taskDurationSeconds": zod.number(),
  "fontSize": zod.enum(['small', 'medium', 'large', 'xlarge']),
  "audioEnabled": zod.boolean(),
  "visualHighlights": zod.boolean(),
  "simplifiedUi": zod.boolean(),
  "breakFrequencyMinutes": zod.number(),
  "colorTheme": zod.enum(['default', 'high_contrast', 'warm', 'cool']),
  "currentDifficulty": zod.number().min(1).max(getAdaptiveSettingsResponseCurrentDifficultyMax)
})


/**
 * @summary Update adaptive settings
 */
export const updateAdaptiveSettingsBodyCurrentDifficultyMax = 5;



export const UpdateAdaptiveSettingsBody = zod.object({
  "taskDurationSeconds": zod.number().optional(),
  "fontSize": zod.enum(['small', 'medium', 'large', 'xlarge']).optional(),
  "audioEnabled": zod.boolean().optional(),
  "visualHighlights": zod.boolean().optional(),
  "simplifiedUi": zod.boolean().optional(),
  "breakFrequencyMinutes": zod.number().optional(),
  "colorTheme": zod.enum(['default', 'high_contrast', 'warm', 'cool']).optional(),
  "currentDifficulty": zod.number().min(1).max(updateAdaptiveSettingsBodyCurrentDifficultyMax).optional()
})

export const updateAdaptiveSettingsResponseCurrentDifficultyMax = 5;



export const UpdateAdaptiveSettingsResponse = zod.object({
  "id": zod.number(),
  "taskDurationSeconds": zod.number(),
  "fontSize": zod.enum(['small', 'medium', 'large', 'xlarge']),
  "audioEnabled": zod.boolean(),
  "visualHighlights": zod.boolean(),
  "simplifiedUi": zod.boolean(),
  "breakFrequencyMinutes": zod.number(),
  "colorTheme": zod.enum(['default', 'high_contrast', 'warm', 'cool']),
  "currentDifficulty": zod.number().min(1).max(updateAdaptiveSettingsResponseCurrentDifficultyMax)
})


/**
 * @summary Get recommended next game based on performance
 */
export const GetGameRecommendationResponse = zod.object({
  "gameType": zod.string(),
  "reason": zod.string(),
  "priority": zod.enum(['high', 'medium', 'low']),
  "suggestedDifficulty": zod.number()
})