# lib/api-zod/src/generated/types/adaptiveSettings.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * CogniLearn Adaptive Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdaptiveSettingsColorTheme } from './adaptiveSettingsColorTheme';
import type { AdaptiveSettingsFontSize } from './adaptiveSettingsFontSize';

export interface AdaptiveSettings {
  id: number;
  taskDurationSeconds: number;
  fontSize: AdaptiveSettingsFontSize;
  audioEnabled: boolean;
  visualHighlights: boolean;
  simplifiedUi: boolean;
  breakFrequencyMinutes: number;
  colorTheme: AdaptiveSettingsColorTheme;
  /**
     * @minimum 1
     * @maximum 5
     */
  currentDifficulty: number;
}